requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
msal==1.17.0
azure-identity==1.7.0